Improve accuracy of anagram generation
This fixes a bug where words that are subgrams of the current word
would not be listed as a target word. That happened because the word
only appeared as a subset of letters that is only contained in subsets
of the current word that have already been checked.

For example, 'believe' has 'lie' as a subgram. But every subset of the
letters b, e, l, i, e, v, e that contains 'lie' just so happens to have
already been checked in words that occur before 'believe' in the search,
like 'live'. So the player did not get credit for inputting 'lie'.

This has slowed down the search process considerably, but it isn't slow
enough to need fixing yet.